requirements.txt
-e git+ssh://git@github.com:bocong/urbandictionary-py.git
discord.py
